requirements.txt
discord.py[voice]~=2.1
requests~=2.28
